+++ client/src/pages/game/[displayPosition]/index.page.tsx
@@ -1,19 +1,6 @@
-/* eslint-disable max-lines */
-//TODO - リファクタリングして行数圧縮
-import {
-  DISPLAY_COUNT,
-  ENEMY_HALF_WIDTH,
-  SCREEN_HEIGHT,
-  SCREEN_WIDTH,
-} from 'commonConstantsWithClient';
-import type { UserId } from 'commonTypesWithClient/branded';
-import type {
-  BulletModelWithPos,
-  EnemyModelWithPos,
-  PlayerModel,
-} from 'commonTypesWithClient/models';
+import { DISPLAY_COUNT, SCREEN_HEIGHT, SCREEN_WIDTH } from 'commonConstantsWithClient';
 import { useRouter } from 'next/router';
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import { Image, Layer, Stage, Text } from 'react-konva';
 import Bomb from 'src/components/Effect/Bomb';
 import Burner from 'src/components/Effect/Burner';
@@ -24,9 +11,9 @@ import { Star } from 'src/components/Effect/Star';
 import { Bullet } from 'src/components/Entity/Bullet';
 import { Enemy } from 'src/components/Entity/Enemy';
 import { Player } from 'src/components/Entity/Player';
-import type { Pos, WindowSize } from 'src/types/types';
+import { useDisplay } from 'src/pages/@hooks/useDisplay';
+import { useWindowSize } from 'src/pages/@hooks/useWindowSize';
 import { staticPath } from 'src/utils/$path';
-import { apiClient } from 'src/utils/apiClient';
 import useImage from 'use-image';
 import styles from './index.module.css';
 
@@ -40,84 +27,15 @@ const Game = () => {
     }
   }
 
-  //ANCHOR - state
-  const [players, setPlayers] = useState<PlayerModel[]>([]);
-  const [enemies, setEnemies] = useState<EnemyModelWithPos[]>([]);
-  const [bullets, setBullets] = useState<BulletModelWithPos[]>([]);
-
-  //TODO: もし、これ以外のエフェクトを追加する場合は、それぞれのエフェクトを区別する型を作成する
-  const [BombEffectPosition, setBombEffectPosition] = useState<Pos[][]>([[]]);
-  const [damagedPlayerIds, setDamagedPlayerIds] = useState<Set<UserId>>(new Set());
-
-  const [windowSize, setWindowSize] = useState<WindowSize>({
-    width: window.innerWidth,
-    height: window.innerHeight,
-  });
-
   const [backgroundImage] = useImage(staticPath.images.space_background_8bit_jpg);
+
+  const { bullets, players, enemies, BombEffectPosition, damagedPlayerIds } = useDisplay({
+    displayPosition,
+  });
 
   const nowTime = Date.now();
 
-  //ANCHOR - check
-  const checkBombEffect = (resEnemies: EnemyModelWithPos[]) => {
-    const currentEnemyIds = new Set(resEnemies.map((e) => e.id));
-    const killedEnemies = enemies.filter(
-      (enemy) =>
-        !currentEnemyIds.has(enemy.id) &&
-        (enemy.pos.x > 1920 * (displayPosition ?? 0) ||
-          enemy.pos.x < 1920 * ((displayPosition ?? 0) + 1))
-    );
-
-    const newEffectPosition = killedEnemies.map((enemy) => {
-      return { x: enemy.pos.x - ENEMY_HALF_WIDTH, y: enemy.pos.y - ENEMY_HALF_WIDTH };
-    });
-
-    setBombEffectPosition((prev) => [...prev.slice(-10), newEffectPosition]);
-  };
-
-  const checkDamageEffect = (resPlayers: PlayerModel[]) => {
-    const damagedPlayers = resPlayers.filter((resPlayer) => {
-      const prevPlayer = players.find((player) => player.id === resPlayer.id);
-      return prevPlayer !== undefined && prevPlayer.score > resPlayer.score;
-    });
-
-    const newDamagedPlayerIds = new Set(damagedPlayers.map((player) => player.id));
-
-    setDamagedPlayerIds(newDamagedPlayerIds);
-  };
-
-  //ANCHOR - fetch
-  const fetchEntities = async () => {
-    const res = await apiClient.entity.$get({
-      query: { displayNumber: Number(displayPosition) },
-    });
-    checkBombEffect(res.enemies);
-    checkDamageEffect(res.players);
-
-    setPlayers(res.players);
-    setEnemies(res.enemies);
-    setBullets(res.bullets);
-  };
-
-  //ANCHOR - effect
-  useEffect(() => {
-    const cancelId = requestAnimationFrame(async () => {
-      await fetchEntities();
-    });
-    return () => cancelAnimationFrame(cancelId);
-  });
-
-  useEffect(() => {
-    const set = () => {
-      setWindowSize({
-        width: window.innerWidth,
-        height: window.innerHeight,
-      });
-    };
-
-    window.addEventListener('resize', set);
-    return () => window.removeEventListener('resize', set);
-  }, []);
+  const { width, height } = useWindowSize();
 
   useEffect(() => {
     const redirectToLobby = async () => {
@@ -136,9 +54,9 @@ const Game = () => {
         height={SCREEN_HEIGHT}
         style={{
           transform: `
-              scale(${windowSize.width / SCREEN_WIDTH}, ${windowSize.height / SCREEN_HEIGHT})
-              translateX(${(windowSize.width - SCREEN_WIDTH) / 2}px)
-              translateY(${(windowSize.height - SCREEN_HEIGHT) / 2}px)
+              scale(${width / SCREEN_WIDTH}, ${height / SCREEN_HEIGHT})
+              translateX(${(width - SCREEN_WIDTH) / 2}px)
+              translateY(${(height - SCREEN_HEIGHT) / 2}px)
               `,
         }}
       >
